Change item.isResearch to eval true if electronic
The DiscoveryStorePoster sometimes generates electronic items out of bib
marc fields. When that happens and there are no other physical items
attached to the bib, the result is that the bib.isResearch method will
return false because the bib has more than 0 items and none of them are
explicitly "Research" (normally evaluated via location and itype). This
change adds an additional clause to cause item.isResearch to evaluate to
true if the item is an electronic item. The anticipated effect is that
some bibs that currently are evaluating as non-research because they
have a single electronic item will now evaluate as research.

Related: https://jira.nypl.org/browse/SCC-283