Add updated mappings for recently added ES props
This updates the section in lib/index.js that establishes the baseline
resources index mapping. This is only used when building a brand new
index, but serves secondarily as documentation of the ES schema. Several
new properties have been added lately without updating this mapping.
This commit corrects that as well as cleans it up with fresh
inline documentation.