Change how items are sorted by box/v. in shelfmark
This changes how the sortable version of the shelfmark is built to
ensure that box/vol/reel numbers occuring mid-string are
'0'-padded. It looks for labels (e.g. box, v., r.) case-insensitively
and ensures differently-cased labels are compared as lowercase (so that
'Box 1' doesn't follow 'box 2', for example). Adds tests.

https://jira.nypl.org/browse/SCC-263