Fix bug where deleted records not deleted from ES
This fixes a bug where a record that fails to be returned from the
DiscoveryStore doesn't result in the related document being deleted.
Previously, SQL lookups that failed were quietly ignored because the id
was assumed to be invalid (i.e. testing). It appears that legitimate
sierra deletions leave a set of statements in the DiscoveryStore that
isn't sufficient for retrieval. (In particular, upon deletion, the
'rdfs:type' statement is deleted because that statement can't be derived
from a `"deleted": true` record. When the DiscoveryApiIndexer
subsequently fetches statements for that record, the query used
*requires* 'rdfs:type', which produces a lookup failure.)

The solution here is to treat all lookup failures as occassions to
delete the corresponding ES document - even if this means we're sending
DELETE queryies against document ids that don't exist (i.e. invalid ids,
ids used for testing). Any id that fails to yeild a minimal set of
statements from the DiscoveryStore should not have a corresponding
document in ES.