SCC-2207 Add Holding to ElasticSearch index
This extends the api-indexer to include `Holding` records in the ElasticSearch index for the SCC API. These records are treated in largely the same way that `Item` records are handled, so it is mostly an extension along those principles. This has been tested (in a limited way) locally and verified at least to write the proper objects to the ES index, though more thorough testing will need to be conducted in the QA environment to ensure that this handles everything properly.

The major changes are:

- A refactor of the `serializer` code. With the addition of an additional class it felt like having all of these in a single file was getting unweildy, so I split them out into their own files in a sub-directory. I think this makes these components easier to maintain/understand
- Added the `resource-holding-serializer` as part of the above work, which is a simple serializer and does not introduce any new ideas into this module
- Extended the `resource-item-serializer` to provide additional fields for separated `shelfMark` data
- Extend the ElasticSearch mapping in `index.js` to add all necessary fields
- Bumped some package versions for NYPL packages to pull in `Holding` updates from upstream, mostly dealing with retrieving these records from discovery-store
- Added some simple tests to validate the serializations