Add encrypted platform creds for aeon links
Changes the platform api key and secret to be encrypted so that we can
store them in public config files. This required also changing
lib/platform-api-client to handle the decryption. The process for
building a platform api client is now async (because on the first pass
we need to decrypt credentials) so this also includes light changes to
the item serializer.