Fix misnamed parallelPublisherLiteral
Fix: Change parallelPublisher to parallelPublisherLiteral so that it
agrees with it's referrent: publisherLiteral. This change has already
been applied to QA and Prod ES indexes. This is just a formality.

Also addresses minor bug in the mapping-check script, causing `object`
type fields to be reported as having different config locally vs remote.